Upgrade to support Neo4j 4.4 and Python 3.10

+++ neomodel/core.py
@@ -57,6 +57,14 @@ def drop_indexes(quiet=True, stdout=None):
             stdout.write(' - Dropping index on label {0} with property {1}.\n'.format(
                 match.group(1), match.group(2)))
         else:
+            # Neo4j 4.3 introduced token lookup indexes
+            # Two are created automatically so should not be dropped
+            # They can be recognized because their labelsOrTypes and properties arrays are empty
+            if not index[7]:
+                if index[8]:
+                    raise ValueError('Index {0} has no labels but has properties({1}). Unknown index'.format(
+                        index[1], index[8]))
+                continue
             db.cypher_query('DROP INDEX ' + index[1])
             stdout.write(' - Dropping index on label {0} with property {1}.\n'.format(
                 index[7][0], index[8][0]))
